pythonanywhere/requirements.txt
dash==2.5.1
dash_auth==1.3.2
dash_bootstrap_components==1.0.0
dash_extensions==0.0.65
dash_iconify==0.1.2
dash_leaflet==0.1.23
dash_mantine_components==0.10.2
numpy==1.19.1
pandas==1.1.0
plotly==5.4.0
python-dotenv==0.21.0
